+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # 1.5.0
 
+## Features
+
+  - New `FormBuilder` class allows easier currying of fields.
+
 ## Breaking changes
 
+  - PHP 5.6 or later is now required.
   - Select API change: prototype more similar to other fields, replaced
     `FormSelect` with `FormOptions`.
   - Button API change: $key removed from prototype.